Refactor shared utils into a util package
