+++ pkg/plugins/platform/utils.go
@@ -2,27 +2,23 @@ package platform
 
 import (
 	"bytes"
-	"errors"
 	"fmt"
 	"html/template"
 	"os"
 	"path"
 	"strings"
-	"time"
 
 	log "github.com/Sirupsen/logrus"
 
+	"errors"
 	"github.com/docker/docker/pkg/homedir"
 	"github.com/drud/ddev/pkg/util"
 	"github.com/drud/ddev/pkg/version"
-	"github.com/drud/drud-go/utils/dockerutil"
 	"github.com/drud/drud-go/utils/system"
-	"github.com/drud/drud-go/utils/try"
 	"github.com/fsouza/go-dockerclient"
-	"github.com/gosuri/uitable"
 )
 
-// PrepLocalSiteDirs creates a site's directories for local dev in ~/.drud/client/site
+// PrepLocalSiteDirs creates a site's directories for local dev in .ddev
 func PrepLocalSiteDirs(base string) error {
 	dirs := []string{
 		".ddev",
@@ -41,67 +37,7 @@ func PrepLocalSiteDirs(base string) error {
 	return nil
 }
 
-// GetPort determines and returns the public port for a given container.
-func GetPort(name string) (int64, error) {
-	client, _ := GetDockerClient()
-	var publicPort int64
-
-	containers, err := client.ListContainers(docker.ListContainersOptions{All: false})
-	if err != nil {
-		return publicPort, err
-	}
-
-	for _, ctr := range containers {
-		if strings.Contains(ctr.Names[0][1:], name) {
-			for _, port := range ctr.Ports {
-				if port.PublicPort != 0 {
-					publicPort = port.PublicPort
-					return publicPort, nil
-				}
-			}
-		}
-	}
-	return publicPort, fmt.Errorf("%s container not ready", name)
-}
-
-// GetPodPort provides a wait loop to help in successfully returning the public port for a given container.
-func GetPodPort(name string) (int64, error) {
-	var publicPort int64
-
-	err := try.Do(func(attempt int) (bool, error) {
-		var err error
-		publicPort, err = GetPort(name)
-		if err != nil {
-			time.Sleep(2 * time.Second) // wait a couple seconds
-		}
-		return attempt < 70, err
-	})
-	if err != nil {
-		return publicPort, err
-	}
-
-	return publicPort, nil
-}
-
-// GetDockerClient returns a docker client for a docker-machine.
-func GetDockerClient() (*docker.Client, error) {
-	// Create a new docker client talking to the default docker-machine.
-	client, err := docker.NewClient("unix:///var/run/docker.sock")
-	if err != nil {
-		log.Fatal(err)
-	}
-	return client, err
-}
-
-// FormatPlural is a simple wrapper which returns different strings based on the count value.
-func FormatPlural(count int, single string, plural string) string {
-	if count == 1 {
-		return single
-	}
-	return plural
-}
-
-// SiteList will prepare and render a list of drud sites running locally.
+// SiteList will prepare and render a list of ddev sites running locally.
 func SiteList(containers []docker.APIContainers) error {
 	apps := map[string]map[string]map[string]string{}
 	var appsFound bool
@@ -114,7 +50,7 @@ func SiteList(containers []docker.APIContainers) error {
 		for _, containerName := range container.Names {
 			for pName, plugin := range PluginMap {
 				if strings.HasPrefix(containerName[1:], plugin.ContainerPrefix()) {
-					ProcessContainer(apps[pName], pName, containerName[1:], container)
+					util.ProcessContainer(apps[pName], pName, containerName[1:], container)
 					break
 				}
 			}
@@ -123,7 +59,7 @@ func SiteList(containers []docker.APIContainers) error {
 
 	if len(apps) > 0 {
 		for k, v := range apps {
-			RenderAppTable(v, k)
+			util.RenderAppTable(v, k)
 		}
 	}
 
@@ -138,95 +74,6 @@ func SiteList(containers []docker.APIContainers) error {
 	}
 
 	return nil
-}
-
-// RenderAppTable will format a table for user display based on a list of apps.
-func RenderAppTable(apps map[string]map[string]string, name string) {
-	if len(apps) > 0 {
-		fmt.Printf("%v %s %v found.\n", len(apps), name, FormatPlural(len(apps), "site", "sites"))
-		table := uitable.New()
-		table.MaxColWidth = 200
-		table.AddRow("NAME", "TYPE", "URL", "DATABASE URL", "STATUS")
-
-		for _, site := range apps {
-			table.AddRow(
-				site["name"],
-				site["type"],
-				site["url"],
-				fmt.Sprintf("127.0.0.1:%s", site["DbPublicPort"]),
-				site["status"],
-			)
-		}
-		fmt.Println(table)
-	}
-
-}
-
-// ProcessContainer will process a docker container for an app listing.
-// Since apps contain multiple containers, ProcessContainer will be called once per container.
-func ProcessContainer(l map[string]map[string]string, plugin string, containerName string, container docker.APIContainers) {
-	label := container.Labels
-	appName := label["com.ddev.site-name"]
-	appType := label["com.ddev.app-type"]
-	containerType := label["com.ddev.container-type"]
-	appRoot := label["com.ddev.approot"]
-	url := label["com.ddev.app-url"]
-
-	_, exists := l[appName]
-	if exists == false {
-		l[appName] = map[string]string{
-			"name":    appName,
-			"status":  container.State,
-			"url":     url,
-			"type":    appType,
-			"approot": appRoot,
-		}
-	}
-
-	var publicPort int64
-	for _, port := range container.Ports {
-		if port.PublicPort != 0 {
-			publicPort = port.PublicPort
-		}
-	}
-
-	if containerType == "web" {
-		l[appName]["WebPublicPort"] = fmt.Sprintf("%d", publicPort)
-	}
-
-	if containerType == "db" {
-		l[appName]["DbPublicPort"] = fmt.Sprintf("%d", publicPort)
-	}
-
-}
-
-// DetermineAppType uses some predetermined file checks to determine if a local app
-// is of any of the known types
-func DetermineAppType(basePath string) (string, error) {
-	defaultLocations := map[string]string{
-		"docroot/scripts/drupal.sh":      "drupal",
-		"docroot/core/scripts/drupal.sh": "drupal8",
-		"docroot/wp":                     "wp",
-	}
-
-	for k, v := range defaultLocations {
-		if FileExists(path.Join(basePath, k)) {
-			return v, nil
-		}
-	}
-
-	return "", fmt.Errorf("unable to determine the application type")
-}
-
-// FileExists checks a file's existence
-// @todo replace this with drud-go/utils version when merged
-func FileExists(name string) bool {
-	if _, err := os.Stat(name); err != nil {
-		if os.IsNotExist(err) {
-			return false
-		}
-	}
-	return true
 }
 
 // EnsureDockerRouter ensures the router is running.
@@ -267,20 +114,6 @@ func EnsureDockerRouter() {
 	if err != nil {
 		fmt.Println(fmt.Errorf("%s - %s", err.Error(), string(out)))
 	}
-
-}
-
-// SubTag replaces current tag on an image or adds one if one does not exist
-func SubTag(image string, tag string) string {
-	if strings.HasSuffix(image, ":"+tag) {
-		return image
-	}
-	if !strings.Contains(image, ":") || (strings.HasPrefix(image, "http") && strings.Count(image, ":") == 1) {
-		return image + ":" + tag
-	}
-	parts := strings.Split(image, ":")
-	parts[len(parts)-1] = tag
-	return strings.Join(parts, ":")
 }
 
 // ComposeFileExists determines if a docker-compose.yml exists for a given app.
@@ -291,9 +124,9 @@ func ComposeFileExists(app App) bool {
 	return true
 }
 
-// Cleanup will clean up legacy apps even if the composer file has been deleted.
+// Cleanup will clean up ddev apps even if the composer file has been deleted.
 func Cleanup(app App) error {
-	client, _ := GetDockerClient()
+	client, _ := util.GetDockerClient()
 
 	containers, err := client.ListContainers(docker.ListContainersOptions{All: false})
 	if err != nil {
@@ -317,7 +150,7 @@ func Cleanup(app App) error {
 	return nil
 }
 
-// CheckForConf checks for a ddev.yaml at the cwd or parent dirs.
+// CheckForConf checks for a config.yaml at the cwd or parent dirs.
 func CheckForConf(confPath string) (string, error) {
 	if system.FileExists(confPath + "/.ddev/config.yaml") {
 		return confPath, nil
@@ -334,65 +167,20 @@ func CheckForConf(confPath string) (string, error) {
 	return "", errors.New("no .ddev/config.yaml file was found in this directory or any parent")
 }
 
-// NetExists checks to see if the docker network for DRUD local development exists.
-func NetExists(client *docker.Client, name string) bool {
-	nets, _ := client.ListNetworks()
-	for _, n := range nets {
-		if n.Name == name {
-			return true
-		}
-	}
-	return false
-}
-
-// EnsureNetwork will ensure the docker network for DRUD local development is created.
-func EnsureNetwork(client *docker.Client, name string) error {
-	if !NetExists(client, name) {
-		netOptions := docker.CreateNetworkOptions{
-			Name:     name,
-			Driver:   "bridge",
-			Internal: false,
-		}
-		_, err := client.CreateNetwork(netOptions)
-		if err != nil {
-			return err
-		}
-		log.Println("Network", name, "created")
-
-	}
-	return nil
-}
-
-// FindContainerByLabels takes a map of label names and values and returns any docker containers which match all labels.
-func FindContainerByLabels(labels map[string]string) (docker.APIContainers, error) {
-	client, _ := dockerutil.GetDockerClient()
-	containers, _ := client.ListContainers(docker.ListContainersOptions{All: true})
-
-	if len(labels) < 1 {
-		return docker.APIContainers{}, fmt.Errorf("the provided list of labels was empty")
+// DetermineAppType uses some predetermined file checks to determine if a local app
+// is of any of the known types
+func DetermineAppType(basePath string) (string, error) {
+	defaultLocations := map[string]string{
+		"docroot/scripts/drupal.sh":      "drupal",
+		"docroot/core/scripts/drupal.sh": "drupal8",
+		"docroot/wp":                     "wp",
 	}
 
-	// First, ensure a site name is set and matches the current application.
-	for i := range containers {
-		matched := true
-		for matchName, matchValue := range labels {
-			// If the label exists check the value to ensure a match
-			if labelValue, ok := containers[i].Labels[matchName]; ok {
-				if labelValue != matchValue {
-					matched = false
-					break
-				}
-			} else {
-				// If the label does not exist, we can just fail immediately.
-				matched = false
-				break
-			}
-		}
-
-		if matched {
-			return containers[i], nil
+	for k, v := range defaultLocations {
+		if system.FileExists(path.Join(basePath, k)) {
+			return v, nil
 		}
 	}
 
-	return docker.APIContainers{}, fmt.Errorf("could not find containers which matched search criteria: %+v", labels)
+	return "", fmt.Errorf("unable to determine the application type")
 }
